Updating getValency() function in ChemElement
1) Removing comments around getValency function
2) valency attribute is now an integer, not a double

+++ source/RMG/jing/chem/ChemElement.java
@@ -135,11 +135,8 @@ public class ChemElement {
       //#]
   }
 
-  //## operation getValency()
-  public double getValency() {
-      //#[ operation getValency()
+  public int getValency() {
       return valency;
-      //#]
   }
 
   //## operation getWeight()
